+++ src/System.IO.Compression.ZipFile/src/System/IO/Compression/ZipFile.Utils.cs
@@ -1,0 +1,80 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Buffers;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace System.IO.Compression
+{
+    internal static partial class ZipFileUtils
+    {
+        // Per the .ZIP File Format Specification 4.4.17.1 all slashes should be forward slashes
+        public const char PathSeparator = '/';
+
+        public static string EntryFromPath(string entry, int offset, int length, ref char[] buffer, bool appendPathSeparator = false)
+        {
+            Debug.Assert(length <= entry.Length - offset);
+            Debug.Assert(buffer != null);
+
+            // Remove any leading slashes from the entry name:
+            while (length > 0)
+            {
+                if (entry[offset] != Path.DirectorySeparatorChar &&
+                    entry[offset] != Path.AltDirectorySeparatorChar)
+                    break;
+
+                offset++;
+                length--;
+            }
+
+            if (length == 0)
+                return appendPathSeparator ? PathSeparator.ToString() : string.Empty;
+
+            int resultLength = appendPathSeparator ? length + 1 : length;
+            EnsureCapacity(ref buffer, resultLength);
+            entry.CopyTo(offset, buffer, 0, length);
+
+            // '/' is a more broadly recognized directory separator on all platforms (eg: mac, linux)
+            // We don't use Path.DirectorySeparatorChar or AltDirectorySeparatorChar because this is
+            // explicitly trying to standardize to '/'
+            for (int i = 0; i < length; i++)
+            {
+                char ch = buffer[i];
+                if (ch == Path.DirectorySeparatorChar || ch == Path.AltDirectorySeparatorChar)
+                    buffer[i] = PathSeparator;
+            }
+
+            if (appendPathSeparator)
+                buffer[length] = PathSeparator;
+
+            return new string(buffer, 0, resultLength);
+        }
+
+        public static void EnsureCapacity(ref char[] buffer, int min)
+        {
+            Debug.Assert(buffer != null);
+            Debug.Assert(min > 0);
+
+            if (buffer.Length < min)
+            {
+                int newCapacity = buffer.Length * 2;
+                if (newCapacity < min)
+                    newCapacity = min;
+                ArrayPool<char>.Shared.Return(buffer);
+                buffer = ArrayPool<char>.Shared.Rent(newCapacity);
+            }
+        }
+
+        public static bool IsDirEmpty(DirectoryInfo possiblyEmptyDir)
+        {
+            using (IEnumerator<string> enumerator = Directory.EnumerateFileSystemEntries(possiblyEmptyDir.FullName).GetEnumerator())
+                return !enumerator.MoveNext();
+        }
+    }
+}
